breaking out KifiVersion to its own file
